refactor: change some decorator methods path
